auto test

+++ Generator.h
@@ -8,12 +8,12 @@ class IntegerGenerator {
 private:
     std::random_device rd;  
     std::mt19937 gen;      
-    std::uniform_int_distribution<int> dis; 
+    std::uniform_int_distribution<int64_t> dis; 
 
 public:
     IntegerGenerator(size_t min, size_t max) : gen(rd()), dis(min, max) {}
 
-    int generate() { return dis(gen); }
+    int64_t generate() { return dis(gen); }
  
 };
